+++ tests/_async/test_http2.py
@@ -53,6 +53,40 @@ async def test_http2_connection():
 
 
 @pytest.mark.anyio
+async def test_http2_connection_closed():
+    origin = Origin(b"https", b"example.com", 443)
+    stream = AsyncMockStream(
+        [
+            hyperframe.frame.SettingsFrame().serialize(),
+            hyperframe.frame.HeadersFrame(
+                stream_id=1,
+                data=hpack.Encoder().encode(
+                    [
+                        (b":status", b"200"),
+                        (b"content-type", b"plain/text"),
+                    ]
+                ),
+                flags=["END_HEADERS"],
+            ).serialize(),
+            hyperframe.frame.DataFrame(
+                stream_id=1, data=b"Hello, world!", flags=["END_STREAM"]
+            ).serialize(),
+            # Connection is closed after the first response
+            hyperframe.frame.GoAwayFrame(stream_id=0, error_code=0).serialize(),
+        ]
+    )
+    async with AsyncHTTP2Connection(
+        origin=origin, stream=stream, keepalive_expiry=5.0
+    ) as conn:
+        await conn.request("GET", "https://example.com/")
+
+        with pytest.raises(RemoteProtocolError):
+            await conn.request("GET", "https://example.com/")
+
+        assert not conn.is_available()
+
+
+@pytest.mark.anyio
 async def test_http2_connection_post_request():
     origin = Origin(b"https", b"example.com", 443)
     stream = AsyncMockStream(
@@ -146,8 +180,8 @@ async def test_http2_connection_with_rst_stream():
 @pytest.mark.anyio
 async def test_http2_connection_with_goaway():
     """
-    If a stream reset occurs, then no response will be returned,
-    but the connection will remain reusable for other requests.
+    If a GoAway frame occurs, then no response will be returned,
+    and the connection will not be reusable for other requests.
     """
     origin = Origin(b"https", b"example.com", 443)
     stream = AsyncMockStream(
@@ -294,3 +328,55 @@ async def test_http2_request_to_incorrect_origin():
     async with AsyncHTTP2Connection(origin=origin, stream=stream) as conn:
         with pytest.raises(RuntimeError):
             await conn.request("GET", "https://other.com/")
+
+
+@pytest.mark.anyio
+async def test_http2_remote_max_streams_update():
+    """
+    If the remote server updates the maximum concurrent streams value, we should
+    be adjusting how many streams we will allow.
+    """
+    origin = Origin(b"https", b"example.com", 443)
+    stream = AsyncMockStream(
+        [
+            hyperframe.frame.SettingsFrame(
+                settings={hyperframe.frame.SettingsFrame.MAX_CONCURRENT_STREAMS: 1000}
+            ).serialize(),
+            hyperframe.frame.HeadersFrame(
+                stream_id=1,
+                data=hpack.Encoder().encode(
+                    [
+                        (b":status", b"200"),
+                        (b"content-type", b"plain/text"),
+                    ]
+                ),
+                flags=["END_HEADERS"],
+            ).serialize(),
+            hyperframe.frame.DataFrame(stream_id=1, data=b"Hello, world!").serialize(),
+            hyperframe.frame.SettingsFrame(
+                settings={hyperframe.frame.SettingsFrame.MAX_CONCURRENT_STREAMS: 50}
+            ).serialize(),
+            hyperframe.frame.DataFrame(
+                stream_id=1, data=b"Hello, world...again!", flags=["END_STREAM"]
+            ).serialize(),
+        ]
+    )
+    async with AsyncHTTP2Connection(origin=origin, stream=stream) as conn:
+        async with conn.stream("GET", "https://example.com/") as response:
+            i = 0
+            async for chunk in response.aiter_stream():
+                if i == 0:
+                    assert chunk == b"Hello, world!"
+                    assert conn._h2_state.remote_settings.max_concurrent_streams == 1000
+                    assert conn._max_streams == min(
+                        conn._h2_state.remote_settings.max_concurrent_streams,
+                        conn._h2_state.local_settings.max_concurrent_streams,
+                    )
+                elif i == 1:
+                    assert chunk == b"Hello, world...again!"
+                    assert conn._h2_state.remote_settings.max_concurrent_streams == 50
+                    assert conn._max_streams == min(
+                        conn._h2_state.remote_settings.max_concurrent_streams,
+                        conn._h2_state.local_settings.max_concurrent_streams,
+                    )
+                i += 1