+++ tests/_sync/test_connection_pool.py
@@ -1,3 +1,4 @@
+import logging
 from typing import List, Optional
 
 import pytest
@@ -157,6 +158,74 @@ def test_trace_request():
         "http11.receive_response_body.complete",
         "http11.response_closed.started",
         "http11.response_closed.complete",
+    ]
+
+
+
+def test_debug_request(caplog):
+    """
+    The 'trace' request extension allows for a callback function to inspect the
+    internal events that occur while sending a request.
+    """
+    caplog.set_level(logging.DEBUG)
+
+    network_backend = MockBackend(
+        [
+            b"HTTP/1.1 200 OK\r\n",
+            b"Content-Type: plain/text\r\n",
+            b"Content-Length: 13\r\n",
+            b"\r\n",
+            b"Hello, world!",
+        ]
+    )
+
+    with ConnectionPool(network_backend=network_backend) as pool:
+        pool.request("GET", "http://example.com/")
+
+    assert caplog.record_tuples == [
+        (
+            "httpcore.connection",
+            logging.DEBUG,
+            "connect_tcp.started host='example.com' port=80 local_address=None timeout=None",
+        ),
+        (
+            "httpcore.connection",
+            logging.DEBUG,
+            "connect_tcp.complete return_value=<httpcore.MockStream>",
+        ),
+        (
+            "httpcore.http11",
+            logging.DEBUG,
+            "send_request_headers.started request=<Request [b'GET']>",
+        ),
+        ("httpcore.http11", logging.DEBUG, "send_request_headers.complete"),
+        (
+            "httpcore.http11",
+            logging.DEBUG,
+            "send_request_body.started request=<Request [b'GET']>",
+        ),
+        ("httpcore.http11", logging.DEBUG, "send_request_body.complete"),
+        (
+            "httpcore.http11",
+            logging.DEBUG,
+            "receive_response_headers.started request=<Request [b'GET']>",
+        ),
+        (
+            "httpcore.http11",
+            logging.DEBUG,
+            "receive_response_headers.complete return_value="
+            "(b'HTTP/1.1', 200, b'OK', [(b'Content-Type', b'plain/text'), (b'Content-Length', b'13')])",
+        ),
+        (
+            "httpcore.http11",
+            logging.DEBUG,
+            "receive_response_body.started request=<Request [b'GET']>",
+        ),
+        ("httpcore.http11", logging.DEBUG, "receive_response_body.complete"),
+        ("httpcore.http11", logging.DEBUG, "response_closed.started"),
+        ("httpcore.http11", logging.DEBUG, "response_closed.complete"),
+        ("httpcore.connection", logging.DEBUG, "close.started"),
+        ("httpcore.connection", logging.DEBUG, "close.complete"),
     ]
 
 
@@ -500,6 +569,68 @@ def test_connection_pool_timeout():
             with pytest.raises(PoolTimeout):
                 extensions = {"timeout": {"pool": 0.0001}}
                 pool.request("GET", "https://example.com/", extensions=extensions)
+
+
+
+def test_connection_pool_timeout_zero():
+    """
+    A pool timeout of 0 shouldn't raise a PoolTimeout if there's
+    no need to wait on a new connection.
+    """
+    network_backend = MockBackend(
+        [
+            b"HTTP/1.1 200 OK\r\n",
+            b"Content-Type: plain/text\r\n",
+            b"Content-Length: 13\r\n",
+            b"\r\n",
+            b"Hello, world!",
+            b"HTTP/1.1 200 OK\r\n",
+            b"Content-Type: plain/text\r\n",
+            b"Content-Length: 13\r\n",
+            b"\r\n",
+            b"Hello, world!",
+        ]
+    )
+
+    # Use a pool timeout of zero.
+    extensions = {"timeout": {"pool": 0}}
+
+    # A connection pool configured to allow only one connection at a time.
+    with ConnectionPool(
+        network_backend=network_backend, max_connections=1
+    ) as pool:
+        # Two consecutive requests with a pool timeout of zero.
+        # Both succeed without raising a timeout.
+        response = pool.request(
+            "GET", "https://example.com/", extensions=extensions
+        )
+        assert response.status == 200
+        assert response.content == b"Hello, world!"
+
+        response = pool.request(
+            "GET", "https://example.com/", extensions=extensions
+        )
+        assert response.status == 200
+        assert response.content == b"Hello, world!"
+
+    # A connection pool configured to allow only one connection at a time.
+    with ConnectionPool(
+        network_backend=network_backend, max_connections=1
+    ) as pool:
+        # Two concurrent requests with a pool timeout of zero.
+        # Only the first will succeed without raising a timeout.
+        with pool.stream(
+            "GET", "https://example.com/", extensions=extensions
+        ) as response:
+            # The first response hasn't yet completed.
+            with pytest.raises(PoolTimeout):
+                # So a pool timeout occurs.
+                pool.request("GET", "https://example.com/", extensions=extensions)
+            # The first response now completes.
+            response.read()
+
+        assert response.status == 200
+        assert response.content == b"Hello, world!"